backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy>=1.4.4,<2.0.0
alembic==1.13.1
asyncpg==0.29.0
databases[postgresql]==0.8.0

# HTTP client for web search and API calls
httpx==0.25.2
aiohttp==3.9.1

# Google APIs for web search fallback
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AI/ML libraries for knowledge base and processing
openai==1.3.8
anthropic==0.7.8
google-generativeai==0.3.2
sentence-transformers>=2.3.0
numpy==1.24.3
pandas==2.1.4

# Text processing and NLP
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Math-specific libraries
sympy==1.12
scipy==1.11.4
matplotlib==3.8.2

# Vector database for knowledge base
chromadb==0.4.18
faiss-cpu==1.7.4

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.1
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Security and validation
cryptography>=41.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1